+++ src/main/java/usagi/exception/UsagiException.java
@@ -1,3 +1,5 @@
+package usagi.exception;
+
 public class UsagiException extends Exception {
     public UsagiException(String message, Throwable cause) {
         super(message, cause);
