+++ src/main/java/usagi/task/Deadline.java
@@ -1,36 +1,37 @@
+package usagi.task;
+
 /**
- * Represents a Event task with title, completion status a from and a to time.
+ * Represents a Deadline task with title, completion status and a by time.
  * This is a concrete class inherited from the abstract Task class.
  */
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task{
-    protected LocalDateTime from;
-    protected LocalDateTime to;
+public class Deadline extends Task{
+    protected LocalDateTime by;
 
     /**
      * Returns a DateTimeFormatter that takes in a LocalDateTime and formats it into a String.
      */
     private static final DateTimeFormatter UI = DateTimeFormatter.ofPattern("MMM d yyyy HH:mm");
 
-    public Event(String title, boolean done, LocalDateTime from, LocalDateTime to) {
+    public Deadline(String title, boolean done, LocalDateTime by) {
         super(title, done);
-        this.from = from;
-        this.to = to;
+        this.by = by;
     }
 
     @Override
     public String type() {
-        return "E";
+        return "D";
     }
 
     public String[] extra() {
-        return new String[]{from.toString(), to.toString()};
+        return new String[]{by.toString()};
     }
 
     public String toString() {
-        return "[E]" + super.toString() + " (from: " + UI.format(from) + " to: " + UI.format(to) + ")";
+        return "[D]" + super.toString() + " (by: " + UI.format(by) + ")";
     }
 }
